requirements.txt
flask
psycopg2
dotenv
pyjwt
crypto
hashlib
dotenv